Fix bug in block commit
Before commit code assumed that block will be present in the
proposal storage at the moment of commit. That's not correct.
If we're catching up node will receive precommits for commit
and then will start accepting block so it's likely that it won't
have block when we get to commit phase