Make all datastructure's fields strict.
In this commit all the data-datastrures are made strict.
In general it's more safe to keep fields strict for all
datastructures that are not control-structures.
Basically the control structure is a structure that define
evaluation order, can have knots, and basically the one that
has Functor, Foldable, Traversable, Applicative, Alternative, Monad
instances implemented.

So the rule of thumb - if structure doesn't have an instance
for those type-classes - it's a value datastructure and it's
fields should be strict. In general all the fields should be
strict by default unless there is a reason for not doing so.
There are some cases when structure should be lazy, but that
should be explained in comments explicitly.

In general making fields strict gives us two benefits:

   1. It's harder to introduce space leaks, but it's still
      possible to do.

   2. GHC errors out on unset strict fields, but it doesn't
      do that for lazy fields. This increases safety in
      presence of refactors.