Pass full block to callback
for validation and update of user state. This gives full information
to application. Most importantly height and time of block

Changes to callbacks are not complete. In order to accomodate
simultaneous use of in-memory state and persisted state it have to be
revised further. Improving support of changing validator set demands the
same.

Fixes #174