Faster retriveCommitRound
Make retrieving commit round for gossip much faster instead of
fetching complete block/last commit we cache round number in
blockchain table.

This functions is called approximatelhy in half of cases when we
or our peer commit new block. This assumes that every node operates
normally and doesn't lag. Savings are modest but on the other hand
implementation is  very cheap.