Rework type class and monads for querying database
Main motovation is to allow interleaving stores to the primary
database and other effects. This is needed in order to implement
custom storage for parts of user state

This become possible because of subtle change in tansaction handling.
Now we let SQLite handle lock contention (but see #311) which simplify
retry logic considerably. But most important rollback of transaction is
implemented by throwing exception which gives relatively simple and
consistent semantics for monadic stacks.