Parametrize Signed by information about key
Instead of storing fingerprint uncoditionally allow to choose
what information identifies signing key.

End goal is removing fingerprint from CryptoSign and making it
standalone. But that would require specifying hashing algorithm
explicitly and CyrptoSign has no place for that. One way to
fix problem is to avaid using fingerprints in first place and
switch to index of key in validator set. It's more compact which is
plus