Add implementation of special case for PBKDF2
In fact it's copied verbatim from xenochain because it's needed
NOW.

Main problems are PBKDF is a bit tricky to parametrize. In this library
we usually select algorithms using types and there's no obvious type class
for that. HMACs are commonly used for that purpose so we can say "use
HMACS" but we don't have them either.

In any case this function will work as stopgap measure