Remove code for in-base persistemce
This code provided useful insights into desired API structure
for handling of user's state. It's also proof of concept for
possibility of keeping state in the database

However it wasn't used much and API is clunky. Also it is obstacle
in devising more uniform API for user state. Thus is makes sense
to remove it. Code is relatively self contained so it would be
relatively easy tto restore it if desired