Move intiial state from Logic to Genesis
It was made in order to unify BlockchainLogic and AppLogic and it make sense
anyway. Genesis is inital step of fold and logic is fold step.

It also makes possible to create migration based on creating "new chain" when
changing type of transaction/block/crypto algorithm. New TX will still point
back to previous chain. Not like it's completely detached.