Refactor gossip in order to get rid of currentState
Instead of returning we keep return value as return value we track it in the
monad. Change is really questionable since we discard all changes to state after
we call `setFinalState'. On one hand we don't do anything after that. On the
other it's not enforced, on third there aren't many sensible things to do after
calling `setFinalState'

Syntactically it's quite pleasing however. `currentState' everywhere looks
positively gnarly.