Refactor gossip logic
We take advantage of the fact that we have handlers that generate messages but
don't change state: reactions to timeouts. And function which change state but
don't genrate new messages

Overall this change clear up gossip logic considerable. Since concerns become
clearly separated. We know what to do when we generate gossip messages.

Tests are still broken and cleanup is needed