Silence tasty's spam bout threads tragically dying
This is hack to make running testing more pleasant. Tasty insists on printing
exceptions in threads failing due to them. But threads failing due to network
exception is pretty normal and expected thing in our program. So we convert
NetworkError/MockNetError to ThreadKilled so it will dutyfully kill thread.

Loss of exception type is not important since we don't use it anyway