Keep progress: move answer before heaader
We are moving answer to a puzzle (nonce) before header because it makes
PoW function more resistant to selfish-mining and migration of compute power
between cryptocurrencies.

After that it will not be possible to hide block header behind its hash
and mining pool members will be able to check whether they mine in the
cryptocurrency of their choice.