Allow STM operations on consensus
We may need to operate on the StateView of consensus - add transactions,
obtain block to mine, etc. This is mostly side related of actual consensus
operations yet we need them to be in sync with actual consensus.

Thus I added a TVar to transactionally operate over consensus for anyone
who need that.

Again, it is an implementation detail and may be changed in future. Most
probably, it will.